rustdoc-search: address nits
